move print table by name into a function
